move cell graph plot to another file cell graph can be plotted separately rename energyplot.generate_gnuplot_energy_plot to energyplot.generate_gnuplot_energy_template add checking if input file exists

+++ energyplot.py
@@ -1,29 +1,35 @@
 #! /usr/bin/python2
+from os.path import exists
+from os import access, R_OK
 
 def generate_temp_csv(energyfn):
-    outsum = open('.sum.energy.dat', 'w')
-    outdif = open('.dif.energy.dat', 'w')
-    outenr = open('.enr.energy.dat', 'w')
-    outenb = open('.enb.energy.dat', 'w')
+    if exists(energyfn) and access(energyfn, R_OK):
+        outsum = open('.sum.energy.dat', 'w')
+        outdif = open('.dif.energy.dat', 'w')
+        outenr = open('.enr.energy.dat', 'w')
+        outenb = open('.enb.energy.dat', 'w')
 
-    num_lines = sum(1 for line in open('energy.csv'))
-    offset = min(500, num_lines / 2)
+        num_lines = sum(1 for line in open('energy.csv'))
+        offset = min(500, num_lines / 2)
 
-    for i, line in enumerate(open('energy.csv')):
-        line = line.strip().split(',')
+        for i, line in enumerate(open('energy.csv')):
+            line = line.strip().split(',')
 
-        if i > offset:
-            outdif.write(str(i) + ' ' + line[4] + '\n')
+            if i > offset:
+                outdif.write(str(i) + ' ' + line[4] + '\n')
 
-        outsum.write(str(i) + ' ' + line[0] + '\n')
-        outenr.write(str(i) + ' ' + line[1] + '\n')
-        outenb.write(str(i) + ' ' + line[2] + '\n')
+            outsum.write(str(i) + ' ' + line[0] + '\n')
+            outenr.write(str(i) + ' ' + line[1] + '\n')
+            outenb.write(str(i) + ' ' + line[2] + '\n')
 
-    for p in (outsum, outdif, outenr, outenb):
-        p.close()
-    print 'Temp csv\'s generated'
+        for p in (outsum, outdif, outenr, outenb):
+            p.close()
+        print 'Temp csv\'s generated'
+    else:
+        print 'File ' + energyfn + 'is inaccessible. Now exit'
+        exit(1)
 
-def generate_gnuplot_energy_plot():
+def generate_gnuplot_energy_template():
     gpl = open('.energy.gpl', 'w')
     titles = ('{/Symbol S}E', '{/Symbol D}E', 'E_R', 'E_B')
     filenames = ('sume.pdf', 'diffe.pdf', 'er.pdf', 'eb.pdf')
@@ -53,4 +59,4 @@ def generate_gnuplot_energy_plot():
 if __name__ == '__main__':
     energyfn = raw_input('Input energy filename: ')
     generate_temp_csv(energyfn)
-    generate_gnuplot_energy_plot()
+    generate_gnuplot_energy_template()
